+++ src/main/java/com/yahoo/ycsb/measurements/TwoInOneMeasurement.java
@@ -53,9 +53,10 @@ public class TwoInOneMeasurement extends OneMeasurement {
    * Using {@link Recorder} to support concurrent updates to histogram.
    *
    * @see com.yahoo.ycsb.OneMeasurement#measure(int)
+   * @param latencyInMicros
    */
   @Override
-  public void measure(int latencyInMicros) {
+  public void measure(long latencyInMicros) {
     thing1.measure(latencyInMicros);
     thing2.measure(latencyInMicros);
   }
